pyproject.toml
[tool.poetry]
name = "jumys2"
version = "0.1.0"
description = ""
authors = ["Daniyal <daniyal.mambetov.04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.28.0"
django-filter = "^24.3"
django-redis = "^5.4.0"
psycopg2 = "^2.9.10"
celery = "^5.4.0"
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pillow = "^11.0.0"
python-environ = "^0.4.54"
djangorestframework-simplejwt = "^5.3.1"
crispy-bootstrap5 = "^2024.10"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
